requirements.txt
fastmcp>=0.1.0
httpx>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
tenacity>=8.2.0

# LangChain + Azure OpenAI
langchain-openai>=0.1.0
langchain-core>=0.2.0
langchain-mcp-adapters>=0.1.0
langgraph>=0.2.0

# Streamlit UI
streamlit>=1.35.0
nest_asyncio>=1.6.0

# SSE transport
uvicorn>=0.29.0

# Development dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=24.0.0
mypy>=1.8.0
ruff>=0.1.0
